refgenie add fixes; #180
